requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==3.1.0
discord.py==2.3.1
dnspython==2.3.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
pip==23.1.2
pytz-deprecation-shim==0.1.0.post0
pytz==2023.3
setuptools==68.0.0
six==1.16.0
wheel==0.40.0
yarl==1.9.2
sentry-sdk==1.27.0
pydantic~=1.10.9
coloredlogs~=15.0.1
python-dotenv~=1.0.0